+++ src/Hook/ThemeHooks.php
@@ -9,6 +9,7 @@ use Drupal\Core\Entity\Display\EntityViewDisplayInterface;
 use Drupal\Core\Entity\EntityInterface;
 use Drupal\Core\Form\FormBuilderInterface;
 use Drupal\Core\Hook\Attribute\Hook;
+use Drupal\farm_grazing_plan\Form\GrazingPlanEventsForm;
 use Drupal\farm_grazing_plan\Form\GrazingPlanTimelineForm;
 use Drupal\farm_grazing_plan\GrazingPlanInterface;
 
@@ -43,6 +44,9 @@ class ThemeHooks {
 
     // Render the grazing plan timeline.
     $build['grazing_plan_timeline'] = $this->formBuilder->getForm(GrazingPlanTimelineForm::class, $plan);
+
+    // Render the grazing plan form.
+    $build['grazing_plan_form'] = $this->formBuilder->getForm(GrazingPlanEventsForm::class, $plan);
   }
 
   /**
@@ -56,6 +60,7 @@ class ThemeHooks {
       return [
         'top' => [
           'grazing_plan_timeline',
+          'grazing_plan_form',
         ],
       ];
     }